Refactor authenticode methods, incorporate code review feedback
